Move tBTC extension setup to pkg/client/client.go
Now that the tBTC handle and extension interfaces are completely chain-
independent for the caller, a single place can kick off the tBTC
extension irrespective of the underlying chain.
